Move WsChannel* and WsMessage to separate files
